refactor(ui): Finish refactoring of local components
Some local components were still using type imports from the old query
client, so clean these.
